requirements.txt
fastapi
uvicorn
pydantic>=2
jaydebeapi
jpype1
pandas
openpyxl
python-dotenv